code refactoring and styling updates
