+++ src/frontend/context/themeContext.tsx
@@ -1,14 +1,11 @@
 import { createContext, useState, useContext } from 'react';
+import { ThemeProviderProps } from '../utility';
 
 type ThemeObj = {
   theme: String,
   switchTheme: () => void;
 }
 const ThemeContext = createContext<ThemeObj>({ theme: 'light', switchTheme: () => { } });
-
-interface ThemeProviderProps {
-  children?: React.ReactNode;
-}
 
 const ThemeProvider: React.FC<ThemeProviderProps> = ({ children }) => {
   const [theme, setTheme] = useState('light');
