+++ src/frontend/pages/category/index.tsx
@@ -1,48 +1,13 @@
 import './category.css';
-import { useEffect, useState } from 'react';
-import { useNavigate, useParams } from 'react-router-dom';
-import { db, quizCollection } from '../../firebase/firebase';
-import {
-    doc,
-    getDoc
-} from 'firebase/firestore';
-import { HOMEPAGE } from '../../routes';
-
-interface OptionsObject {
-    one: String,
-    two: String,
-    three: String
-}
-interface ElementObject {
-    title: String,
-    description: String,
-    banner: String,
-    questions: Array<String>,
-    answers: Array<String>,
-    options: Array<OptionsObject>
-}
+import { useParams } from 'react-router-dom';
+import { useCategoryId, ElementObject } from '../../utility';
 
 export default function Category() {
     const { categoryId } = useParams();
-    const navigate = useNavigate();
-    const [quizData, setQuizData] = useState<ElementObject[]>([]);
-
-    useEffect(() => {
-        if (categoryId) {
-            const docRef = doc(db, quizCollection, `${categoryId}`);
-            (async function () {
-                const docSnap = await getDoc(docRef);
-                const data = Object.values(docSnap.data() ?? {})
-                setQuizData(data);
-            })()
-        }
-        else {
-            navigate(HOMEPAGE)
-        }
-    }, [categoryId, navigate])
+    const quizData = useCategoryId(`${categoryId}`);
 
     return <div>
-        <h1 className="title categoryTitle xl sb cen sm-s">{categoryId}</h1>
+        <h1 className="title categoryTitle xl sb cen">{categoryId}</h1>
         <h1 className="subtitle lg bd cen">Pick a quiz from these options</h1>
         <div className="category">
             {quizData.map((element: ElementObject) => {
